Moved env's, finished moving types and added httpService and service file
